Pie menu (#1226)
* use pie menu for solfege

* hide div when closing

* fill in center of the menu

* add i18n support to pie menu

* remove extra </div>

* add pie menu support for note name block

* add east indian solfege to pie menu

* fix problem with block value when creating new east indian solfege blocks

* add tooltips to accidents menu items

* add drum pie menu; fix regression in loadSynth

* add pie menu for voice selector

* add pie-menu related libs

* add pitch preview to pie menu

* use block colors

* align selected accidental

* use octave from pitch block if available

* fix some regressions in synth loading; add octave selector

* Add preview sounds for voices

* check for null case

* use color to denote category in pie menus for voice

* add background to pie menu div

* cleaning up unused input code

* fix problem with voice preview

* tweak voice menu parameters

* fix docking problem with value blocks inside notes

* add accidental pie menu; clean up

* invert mode pie menu

* boolean mode pie menu

* don't need exit button when selecting accidentals

* add octave pie menu

* use pitch and octave pie menus wherever applicable

* add pie menus for filter an osc

* add interval pie menu

* interval blocks should match pitch colors

* update icon colors too

* group meter with notes; intervals with pitch

* add mode pie menu

* pie menu for temperament

* add preview support for the intervals

* use same layout for accidentals as pitch

* order notes by fifths in pie menu when setting key

* always update block label and value from pie menu pitch

* remove unused DIVs

* ensure selections are stored immediately

* speed up animation of pie menus

* separate number tabs for each interval name

* unison === perfect 1

* add exit button to octave and accident menus

* play notes in mode selector

* keep mode menu on the screen

* keep all pie menus on the screen

* disable keyboard when pie menu is open

+++ js/blocks.js
@@ -965,11 +965,6 @@ function Blocks () {
             this._deletePitchBlocks(thisBlock);
             return this.blockList[thisBlock].connections[0];
         } else {
-            if (thisBlockobj && thisBlockobj.connections.length === 1) {
-                console.log('Value block encountered? ' + thisBlockobj.name);
-                return;
-            }
-
             while (thisBlockobj.connections[0] != null) {
                 var i = thisBlockobj.connections[0];
                 if (NOTEBLOCKS.indexOf(this.blockList[i].name) !== -1) {
@@ -1364,7 +1359,7 @@ function Blocks () {
 
             // Remove the silence block (if it is present) after
             // adding a new block inside of a note block.
-            if (this._insideNoteBlock(thisBlock) != null) {
+            if (this._insideNoteBlock(thisBlock) != null && this.blockList[thisBlock].connections.length > 1) {
                 // If blocks are inserted above the silence block.
                 if (insertAfterDefault) {
                     newBlock = this.deletePreviousDefault(thisBlock);
@@ -2144,8 +2139,6 @@ function Blocks () {
             postProcessArg = [thisBlock, _('true')];
         } else if (name === 'solfege') {
             postProcessArg = [thisBlock, 'sol'];
-        } else if (name === 'eastindiansolfege') {
-            postProcessArg = [thisBlock, 'pa'];
         } else if (name === 'notename') {
             postProcessArg = [thisBlock, 'G'];
         } else if (name === 'drumname') {
@@ -2156,6 +2149,16 @@ function Blocks () {
             postProcessArg = [thisBlock, DEFAULTOSCILLATORTYPE];
         } else if (name === 'voicename') {
             postProcessArg = [thisBlock, DEFAULTVOICE];
+        } else if (name === 'eastindiansolfege') {
+            postProcess = function (args) {
+                var thisBlock = args[0];
+                var value = args[1];
+                that.blockList[thisBlock].value = value;
+                that.blockList[thisBlock].text.text =  WESTERN2EISOLFEGENAMES[value];
+                that.blockList[thisBlock].container.updateCache();
+            };
+
+            postProcessArg = [thisBlock, 'sol'];
         } else if (name === 'modename') {
             postProcess = function (args) {
                 var thisBlock = args[0];
@@ -2997,6 +3000,47 @@ function Blocks () {
                 }
             } else {
                 return this._insideNoteBlock(cblk);
+            }
+        }
+    };
+
+    this.findPitchOctave = function (blk) {
+        // Returns octave associated with pitch block.
+        if (blk === null) {
+            return 4;
+        }
+
+        if (['pitch', 'setpitchnumberoffset', 'invert1', 'tofrequency'].indexOf(this.blockList[blk].name) !== -1) {
+            var oblk = this.blockList[blk].connections[2];
+            if (oblk === null) {
+                return 4;
+            } else if (this.blockList[oblk].name === 'number') {
+                return this.blockList[oblk].value;
+            } else {
+                return 4;
+            }
+        } else {
+            return 4;
+        }
+    };
+
+    this.setPitchOctave = function (blk, octave) {
+        // Set octave associated with pitch block
+        if (blk === null) {
+            return;
+        }
+
+        if (['pitch', 'setpitchnumberoffset', 'invert1', 'tofrequency'].indexOf(this.blockList[blk].name) !== -1) {
+            var oblk = this.blockList[blk].connections[2];
+            if (oblk !== null && this.blockList[oblk].name === 'number') {
+                var thisBlock = this.blockList[oblk];
+                thisBlock.value = octave;
+                thisBlock.text.text = octave.toString();
+
+                // Make sure text is on top.
+                var z = thisBlock.container.children.length - 1;
+                thisBlock.container.setChildIndex(thisBlock.text, z);
+                thisBlock.container.updateCache();
             }
         }
     };
